// src/modules/config/db/functions/delete/postTerm.d.ts
interface DeletePostTermParamDocument {
    termId: number
}

interface DeletePostTermContentParamDocument {
    termId: number
}

interface DeletePostTermLinkParamDocument {
    postId: number
}

export {
    DeletePostTermParamDocument,
    DeletePostTermContentParamDocument,
    DeletePostTermLinkParamDocument
};